Make configuration descriptor available

+++ java-does-usb/src/main/java/net/codecrete/usb/USBControlTransfer.java
@@ -12,13 +12,17 @@ package net.codecrete.usb;
  * <p>
  * See USB specification for additional information
  * </p>
+ * <p>
+ * For control requests directed to an interface or endpoint,
+ * the lower byte of {@code index} contains the interface or endpoint number.
+ * </p>
  *
- * @param requestType request type
- * @param recipient   recipient
- * @param request     request code
- * @param value       value
- * @param index       index
+ * @param requestType request type (bits 5 and 6 of {@code bmRequestType})
+ * @param recipient   recipient (bits 0–4 of {@code bmRequestType})
+ * @param request     request code ({@code bRequest})
+ * @param value       value ({@code wValue})
+ * @param index       index ({@code wIndex}).
  */
-public record USBControlTransfer(USBRequestType requestType, USBRecipient recipient, byte request, short value,
-                                 short index) {
+public record USBControlTransfer(USBRequestType requestType, USBRecipient recipient, int request, int value,
+                                 int index) {
 }
